moved tests in UI folder
