+++ tests/specs/CP/UI/02-Home.spec.js
@@ -1,19 +1,19 @@
 import { test, expect, chromium } from '@playwright/test';
 import { v4 as uuidv4 } from 'uuid';
-import {loginCS} from '../../helper/CS/functions.js';
-import {DimiDI} from '../../credentials-DI.json';
-import {DimiCS} from '../../credentials-CS.json';
+import {loginCS} from '../../../helper/CS/functions.js';
+import {DimiDI} from '../../../credentials-DI.json';
+import {DimiCS} from '../../../credentials-CS.json';
 import exp from 'constants';
 const fs = require("fs");
 // import TransportsCP from '../../screenObjects/CP/Transports.screen.js';
 
-const LoginScreenCP = require('../../screenObjects/CP/login.screen.js');
-const LoginScreenDI = require('../../screenObjects/English/DI/login.screen.js');
-const NavigationCP = require('../../screenObjects/CP/navigation.wrapper.js');
-const TransportsCP = require('../../screenObjects/CP/Transports.screen.js');
-const TicketCategoriesCP = require('../../screenObjects/CP/TicketCategories.screen.js');
-const UsersCP = require('../../screenObjects/CP/Users.screen.js');
-const HomeScreenDI = require('../../screenObjects/English/DI/home.screen.js');
+const LoginScreenCP = require('../../../screenObjects/CP/login.screen.js');
+const LoginScreenDI = require('../../../screenObjects/English/DI/login.screen.js');
+const NavigationCP = require('../../../screenObjects/CP/navigation.wrapper.js');
+const TransportsCP = require('../../../screenObjects/CP/Transports.screen.js');
+const TicketCategoriesCP = require('../../../screenObjects/CP/TicketCategories.screen.js');
+const UsersCP = require('../../../screenObjects/CP/Users.screen.js');
+const HomeScreenDI = require('../../../screenObjects/English/DI/home.screen.js');
 
 const testUserDI = DimiDI;
 const testUserCS = DimiCS;
